requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchmetrics>=1.4.0
opencv-python>=4.8.0
matplotlib>=3.7.0
tqdm>=4.65.0
numpy>=1.24.0
Pillow>=10.0.0
pytest>=7.4.0
datasets>=2.14.0
deeplake<4
pycocotools
black>=23.0.0
flake8>=6.0.0
timm>=0.9.0